+++ ckb-indexer/src/main/java/org/nervos/indexer/model/ScriptSearchMode.java
@@ -1,0 +1,13 @@
+package org.nervos.indexer.model;
+
+import com.google.gson.annotations.SerializedName;
+
+public enum ScriptSearchMode {
+  // search script with prefix
+  @SerializedName("prefix")
+  Prefix,
+  // search script with exact match
+  @SerializedName("exact")
+  Exact,
+  ;
+}