+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 2.1.1 (2023-03-23)
+- feat: support with_cycles for get_block and get_block_by_number rpc (#623)
+- feat: support packed rpcs (#624)
+- feat: support indexer exact search mod (#627)
+
+## 🚀 Features
 # 2.1.0 (2022-12-26)
 
 ## 🚀 Features